backend/manage.py
#!/usr/bin/env python
"""Django's command-line utility for Playto administrative tasks."""
import os
import sys
from pathlib import Path

from dotenv import load_dotenv


def main():
    """Run Django management commands with local settings by default."""
    # Load backend/.env so PostgreSQL, Redis, and Django config are available
    # without manual shell exports. The .env file is gitignored; .env.example is tracked.
    env_path = Path(__file__).resolve().parent / ".env"
    load_dotenv(env_path)
    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings.local")
    try:
        from django.core.management import execute_from_command_line
    except ImportError as exc:
        raise ImportError(
            "Couldn't import Django. Are you sure it's installed and "
            "available on your PYTHONPATH environment variable? Did you "
            "forget to activate a virtual environment?"
        ) from exc
    execute_from_command_line(sys.argv)


if __name__ == "__main__":
    main()
