requirements.txt
Django==5.2.13
celery==5.6.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.12
python-dotenv==1.1.0
redis==7.4.0
